refactor: Constants
Refactor constants, to take into account tweaks to stripes-erm-components merged along with ERM-2937
