+++ src/components/Amendment/Amendment.js
@@ -32,7 +32,7 @@ import {
 } from '../viewSections';
 
 import { useLicensesContexts } from '../../hooks';
-import { CUSTPROP_ENDPOINT } from '../../constants/endpoints';
+import { CUSTPROP_ENDPOINT } from '../../constants';
 
 const Amendment = ({
   data,
