+++ src/components/LicenseForm/LicenseForm.js
@@ -35,7 +35,7 @@ import {
   FormSupplementaryDocs,
 } from '../formSections';
 
-import { CUSTPROP_ENDPOINT } from '../../constants/endpoints';
+import { CUSTPROP_ENDPOINT } from '../../constants';
 import { useLicensesContexts } from '../../hooks';
 
 const LicenseForm = ({
