--- src/constants/errorTypes.js
@@ -1,4 +1,0 @@
-export default {
-  INVALID_JSON_ERROR: 'invalidJsonError',
-  JSON_ERROR: 'jsonError'
-};
