+++ src/constants/index.js
@@ -1,3 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as errorTypes } from './errorTypes';
 export { default as statuses } from './statuses';
+export * from './endpoints';
