+++ src/routes/CreateAmendmentRoute/CreateAmendmentRoute.js
@@ -8,7 +8,8 @@ import { CalloutContext, useOkapiKy } from '@folio/stripes/core';
 import { getRefdataValuesByDesc } from '@folio/stripes-erm-components';
 
 import Form from '../../components/AmendmentForm';
-import { LICENSE_ENDPOINT } from '../../constants/endpoints';
+
+import { LICENSE_ENDPOINT } from '../../constants';
 import { useLicenseRefdata } from '../../hooks';
 
 const [
